Refactor user services
